+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/intellij/report/MarkdownReportBuilder.kt
@@ -1,8 +1,8 @@
 package io.github.surpsg.deltacoverage.report.intellij.report
 
 import com.intellij.rt.coverage.report.Reporter
-import io.github.surpsg.deltacoverage.config.ReportsConfig
 import io.github.surpsg.deltacoverage.report.ReportBound
+import io.github.surpsg.deltacoverage.report.ReportContext
 import io.github.surpsg.deltacoverage.report.ReportType
 import io.github.surpsg.deltacoverage.report.intellij.coverage.IntellijRawCoverageDataProvider
 import io.github.surpsg.deltacoverage.report.path.ReportPathStrategy
@@ -14,11 +14,14 @@ internal class MarkdownReportBuilder(
     val reportBound: ReportBound,
     private val reportView: String,
     private val reportsConfig: ReportsConfig,
+    private val reportContext: ReportContext,
     private val reporter: Reporter,
 ) : ReportBuilder {
 
     override fun buildReport() {
-        val reportPath: File = ReportPathStrategy.Markdown(reportsConfig).buildReportPath(reportBound)
+        val reportPath: File = ReportPathStrategy.Markdown(reportContext.deltaCoverageConfig.reportsConfig)
+            .buildReportPath(reportBound)
+
         reportPath.outputStream().use { os ->
             val buildContext = BuildContext {
                 viewName = reportView
@@ -26,6 +29,10 @@ internal class MarkdownReportBuilder(
                 reportBound = this@MarkdownReportBuilder.reportBound
                 coverageDataProvider = IntellijRawCoverageDataProvider(reporter.projectData)
                 outputStream = os
+
+                reportContext.deltaCoverageConfig.coverageRulesConfig.entitiesRules.forEach { (entity, ratio) ->
+                    targetCoverage(entity, ratio.minCoverageRatio)
+                }
             }
             TextualReportFacade.generateReport(buildContext)
         }