+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/jacoco/analyzable/AnalyzableReport.kt
@@ -1,11 +1,6 @@
 package io.github.surpsg.deltacoverage.report.jacoco.analyzable
 
-import io.github.surpsg.deltacoverage.config.ReportsConfig
-import io.github.surpsg.deltacoverage.report.EnabledReportTypeFactory
-import io.github.surpsg.deltacoverage.report.ReportBound
 import io.github.surpsg.deltacoverage.report.ReportContext
-import io.github.surpsg.deltacoverage.report.ReportType
-import io.github.surpsg.deltacoverage.report.jacoco.report.JacocoReport
 import io.github.surpsg.deltacoverage.report.jacoco.report.VerifiableReportVisitor
 import org.jacoco.core.analysis.Analyzer
 import org.jacoco.core.analysis.ICoverageVisitor
@@ -20,36 +15,18 @@ internal interface AnalyzableReport {
 internal fun analyzableReportFactory(
     reportContext: ReportContext
 ): Set<AnalyzableReport> {
-    val allReports = mutableSetOf<AnalyzableReport>()
+    return reportFactory(reportContext)
+        .map { reportMode ->
+            when (reportMode) {
+                is JacocoDeltaReport -> DeltaCoverageAnalyzableReport(
+                    reportContext.deltaCoverageConfig.coverageRulesConfig,
+                    reportMode,
+                )
 
-    val reportsConfig: ReportsConfig = reportContext.deltaCoverageConfig.reportsConfig
-    val enabledReportTypes: Set<ReportType> = EnabledReportTypeFactory.obtain(reportsConfig)
-
-    allReports += DeltaCoverageAnalyzableReport(
-        reportContext,
-        ReportBound.DELTA_REPORT.buildJacocoReports(reportsConfig, enabledReportTypes),
-    )
-
-    if (reportsConfig.fullCoverageReport) {
-        allReports += FullCoverageAnalyzableReport(
-            ReportBound.FULL_REPORT.buildJacocoReports(reportsConfig, enabledReportTypes)
-        )
-    }
-
-    return allReports
+                else -> FullCoverageAnalyzableReport(
+                    reportMode,
+                    reportContext.deltaCoverageConfig.coverageRulesConfig,
+                )
+            }
+        }.toSet()
 }
-
-private fun ReportBound.buildJacocoReports(
-    reportsConfig: ReportsConfig,
-    reportTypes: Set<ReportType>
-): List<JacocoReport> {
-    return reportTypes.map { reportType ->
-        JacocoReport(
-            reportType,
-            this,
-            reportsConfig,
-        )
-    }
-}
-
-