+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/jacoco/analyzable/DeltaCoverageAnalyzableReport.kt
@@ -1,5 +1,6 @@
 package io.github.surpsg.deltacoverage.report.jacoco.analyzable
 
+import io.github.surpsg.deltacoverage.config.CoverageRulesConfig
 import io.github.surpsg.deltacoverage.diff.CodeUpdateInfo
 import io.github.surpsg.deltacoverage.diff.parse.ClassFile
 import io.github.surpsg.deltacoverage.report.ReportBound
@@ -13,6 +14,8 @@ import org.jacoco.core.analysis.ICoverageVisitor
 import org.jacoco.core.data.ExecutionDataStore
 
 internal class DeltaCoverageAnalyzableReport(
+    private val coverageRulesConfig: CoverageRulesConfig,
+    private val jacocoDeltaReport: JacocoDeltaReport,
     private val reportContext: ReportContext,
     jacocoReports: List<JacocoReport>,
 ) : FullCoverageAnalyzableReport(jacocoReports) {
@@ -36,6 +39,28 @@ internal class DeltaCoverageAnalyzableReport(
         return FilteringAnalyzer(executionDataStore, coverageVisitor, classFileFilter) {
             ModifiedLinesFilter(codeUpdateInfo)
         }
+    }
+
+    private fun createViolationCheckVisitor(
+        rules: List<Rule>
+    ): IReportVisitor {
+        val violationsOutputResolver = ViolationsOutputResolver(coverageRulesConfig)
+        val coverageViolationsPropagator = CoverageViolationsPropagator()
+
+        class CoverageRulesVisitor(
+            rulesCheckerVisitor: IReportVisitor
+        ) : IReportVisitor by rulesCheckerVisitor {
+
+            override fun visitEnd() {
+                val violations: List<String> = violationsOutputResolver.getViolations()
+                coverageViolationsPropagator.propagate(coverageRulesConfig, violations)
+            }
+        }
+
+        return RulesChecker()
+            .apply { setRules(rules) }
+            .createVisitor(violationsOutputResolver)
+            .let { CoverageRulesVisitor(it) }
     }
 
     override val reportBound: ReportBound = ReportBound.DELTA_REPORT