+++ CHANGELOG.md
@@ -1,10 +1,24 @@
 # Delta-Coverage Gradle plugin Changelog
 
-## <NEXT-RELEASE>
+## 2.5.0
+
+### Fixed
+- #133 Fixed configuration cache compatibility issue.
+
+### Changed
+- Min supported Gradle version is **7.0**.
+
+
+## 2.4.0
 
 ### Changed
 - Source code paths are inferred from sourceSets if no custom paths are specified.
-- `Markdown` report and `Console` report now are printed without `source` column.
+- `Markdown` and `Console` reports:
+  - now don't have `source` column.
+  - now contain min coverage data.
+  - (console) contain `✔`/`✖` regards to coverage check success or failure.
+  - (markdown) contain 🟢/🔴 regards to coverage check success or failure.
+- Console report now enabled by the default.
 
 ### Added
 - Added GitHub Action for posting delta coverage report to PR comment. See [docs](./actions/delta-coverage-report/README.md)